% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{bind_resps}
\alias{bind_resps}
\title{Combine a list of formatted responses}
\usage{
bind_resps(resps)
}
\arguments{
\item{resps}{list of formatted responses}
}
\value{
single formatted response
}
\description{
Combine a list of formatted responses
}
