% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_referenced_tweets}
\alias{flatten_referenced_tweets}
\title{Title}
\usage{
flatten_referenced_tweets(tweets_resp)
}
\arguments{
\item{tweets_resp}{}
}
\value{
formatted response
}
\description{
Title
}
