% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tweets.R
\name{get_tweets}
\alias{get_tweets}
\title{Get tweets for specified screen name}
\usage{
get_tweets(
  bearer_token,
  screen_name,
  n = 100,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{bearer_token}{character string giving the bearer token used for
authorisation.}

\item{screen_name}{character string giving the twitter handle to download
tweets from.}

\item{n}{integer giving the maximum number of tweets to download.}

\item{start_date}{date if specified is the earliest dated tweets to include.}

\item{end_date}{date if specified is the latest dated tweets to include.}
}
\value{
a tibble of the tweet data.
}
\description{
Return a data frame of tweets from a specified screen name. The function
will return at most the last 3,200 tweets (this is a limit of the API) and
these can be limited by either specifying the number of tweets to return, the
time range that tweets fall in, or both. Note that the latest tweets are
returned first. Twitter rate limit the API call to 15 requests every 15
minutes and a maximum of 10,000 tweets per month.
}
\details{
Referenced tweets refers to retweets, retweets with comments, and replies.
The API truncates the text on these tweets to 140 characters and so an
additional \code{rerenced_tweets_text} field is added to the returned tibble.
}
\examples{
get_tweets(
 Sys.getenv("BEARER_TOKEN"),
 "BBCLauraK",
  n = 5,
  start_date = as.Date("2023-07-16") - 4,
  end_date = as.Date("2023-07-16")
)
}
