% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{join_replied_to_usernames}
\alias{join_replied_to_usernames}
\title{Lookup usernames for user ids}
\usage{
join_replied_to_usernames(df, token)
}
\arguments{
\item{df}{data frame of tweets}

\item{token}{character string giving a bearer token used for
authorisation.}
}
\value{
data frame of tweets
}
\description{
Lookup usernames for user ids
}
