% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_search_all.R
\name{tweets_search_all}
\alias{tweets_search_all}
\title{Search of tweets for specified screen name}
\usage{
tweets_search_all(
  token,
  query,
  ...,
  .n = Inf,
  .start_date = NULL,
  .end_date = NULL,
  .format = "parsed"
)
}
\arguments{
\item{token}{character string giving a bearer token used for authorisation.}

\item{query}{character string (max length 500) giving query string.}

\item{...}{values to be passed on the the API}

\item{.n}{optional integer giving the maximum number of tweets to download
(must be at least 5).}

\item{.start_date}{optional date if specified is the earliest dated tweets
to include.}

\item{.end_date}{optional date if specified is the latest dated tweets to
include.}

\item{.format}{optional one of "parsed", "body", or "raw"}
}
\value{
a formatted list of JSONs returned from the API
}
\description{
Return tweets from a specified screen name that match the passed query
string. Note that the maximum length of the query string is unclear but has
been set to 500 here.
}
\details{
Endpoint documentation:
https://developer.twitter.com/en/docs/twitter-api/tweets/search/api-reference/get-tweets-search-all

Query documentation:
https://developer.twitter.com/en/docs/twitter-api/tweets/search/integrate/build-a-query
}
