% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_by_username.R
\name{user_by_username}
\alias{user_by_username}
\title{Get a variety of information about a user specified by their username}
\usage{
user_by_username(token, screen_name, ..., .format = "body")
}
\arguments{
\item{token}{character string giving a bearer token used for
authorisation.}

\item{screen_name}{character string giving the twitter screen name to
download details of.}

\item{...}{values to be passed on the the API}

\item{.format}{optional one of "parsed", "body" (default), or "raw"}
}
\value{
a list of JSONs returned from the API
}
\description{
Endpoint documentation:
https://developer.twitter.com/en/docs/twitter-api/users/lookup/api-reference/get-users-by-username-username
}
