% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_liked_tweets.R
\name{user_liked_tweets}
\alias{user_liked_tweets}
\title{Get tweets a user has liked}
\usage{
user_liked_tweets(token, screen_name, n = 100, ...)
}
\arguments{
\item{token}{optional character string giving a bearer token used for
authorisation.#'}

\item{screen_name}{character string giving the twitter screen name to
download tweets from.}

\item{n}{maximum number of liked tweets to return}

\item{...}{values to be passed on the the API}
}
\value{
a list of JSONs returned from the API
}
\description{
Get tweets a user has liked
}
