% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_referenced_tweets}
\alias{flatten_referenced_tweets}
\title{Collapse down a returned list of tweets}
\usage{
flatten_referenced_tweets(tweets_resp)
}
\arguments{
\item{tweets_resp}{response to be flattened}
}
\value{
data frame of tweets
}
\description{
Flatten a list of tweets by joining on the text of any
referenced tweets and any errors.
}
