% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtwitter-package.R
\docType{package}
\name{readtwitter-package}
\alias{readtwitter}
\alias{readtwitter-package}
\title{readtwitter: Read Tweets From Twitter}
\description{
Package to access and retrieve Tweets via the (paid) Twitter/X API, using either a set number of Tweets or a date range as parameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DataS-DHSC/readtwitter}
  \item Report bugs at \url{https://github.com/DataS-DHSC/readtwitter/issues}
}

}
\author{
\strong{Maintainer}: Matthew Katz \email{matthew.katz@dhsc.gov.uk}

}
\keyword{internal}
