% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readtwitter-package.R
\docType{package}
\name{readtwitter-package}
\alias{readtwitter}
\alias{readtwitter-package}
\title{readtwitter: Collect Twitter data from R}
\description{
readtwitter is a simple wrapper for the Twitter API 2.0 endpoints.
It's goal is to provide a stop gap solution while the rtweet package
is updated to make it compatible with API 2.0. As such it tries to
mimic signatures and functionality as much as possible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DataS-DHSC/readtwitter}
  \item Report bugs at \url{https://github.com/DataS-DHSC/readtwitter/issues}
}

}
\author{
\strong{Maintainer}: Matthew Katz \email{matthew.katz@dhsc.gov.uk}

}
\keyword{internal}
