% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets_search_all.R
\name{tweets_search_keywords}
\alias{tweets_search_keywords}
\title{Title}
\usage{
tweets_search_keywords(
  token,
  screen_name,
  keywords,
  ...,
  include_retweets = TRUE,
  include_replies = TRUE,
  include_quotes = TRUE,
  .n = Inf,
  .start_date = NULL,
  .end_date = NULL,
  .format = "parsed"
)
}
\arguments{
\item{token}{}

\item{screen_name}{}

\item{keywords}{}

\item{...}{}

\item{include_retweets}{}

\item{include_replies}{}

\item{include_quotes}{}

\item{.n}{}

\item{.start_date}{}

\item{.end_date}{}

\item{.format}{}
}
\description{
Title
}
